requirements.txt
opendssdirect.py>=0.7.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
networkx>=2.6.0
openpyxl>=3.0.0
xlrd>=2.0.0
tqdm>=4.62.0
PyQt5>=5.15.0  # For GUI applications if needed
jupyter>=1.0.0  # For notebook development 